CkanLocaliser implementing first rough cut at Validate capability. Currently validates the entire Ckan repository as valid.
Note by design it will fail many many files that are both valid json files and valid ckan files according to the schema.
it will also by design as a feature..., pass lots of files that are invalid by the ckan schema.
If someone wants to validate a file, against the schema they should use tool meant for doing that.
This is not that tool and never will be.